+++ packages/react/src/CuriBase.tsx
@@ -1,9 +1,16 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { CuriRouter, Response, Navigation } from "@curi/core";
-import { CuriContext, CuriProps } from "./interface";
+import {
+  CuriRouter,
+  ResponseHandlerProps,
+  Response,
+  Navigation
+} from "@curi/core";
+import { CuriContext } from "./interface";
 
-export type CuriRenderFn = (props: CuriProps) => React.ReactElement<any>;
+export type CuriRenderFn = (
+  props: ResponseHandlerProps
+) => React.ReactElement<any>;
 
 export interface CuriBaseProps {
   router: CuriRouter;