+++ packages/react/src/index.ts
@@ -1,8 +1,8 @@
 export { ActiveProps } from "./Active";
 export { BlockProps } from "./Block";
 export { CuriBaseProps } from "./CuriBase";
-export { CuriousProps, CuriousRenderProps } from "./Curious";
-export { CuriProps, CuriContext } from "./interface";
+export { CuriousProps } from "./Curious";
+export { CuriContext } from "./interface";
 export { LinkProps, LinkState, ActiveLink } from "./Link";
 export { ResponsiveProps } from "./ResponsiveBase";