+++ packages/react/src/interface.ts
@@ -1,11 +1,5 @@
-import { CuriRouter, Response, Navigation } from "@curi/core";
+import { ResponseHandlerProps } from "@curi/core";
 
 export interface CuriContext {
-  curi: CuriProps;
-}
-
-export interface CuriProps {
-  router: CuriRouter;
-  response: Response;
-  navigation: Navigation;
+  curi: ResponseHandlerProps;
 }