+++ packages/react/types/Curious.d.ts
@@ -2,16 +2,11 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { CuriContext } from "./interface";
-import { CuriRouter, Response, Navigation } from "@curi/core";
+import { CuriRouter, ResponseHandlerProps, Response, Navigation } from "@curi/core";
 export interface CuriousProps {
-    render(p: CuriousRenderProps): React.ReactElement<any>;
+    render(p: ResponseHandlerProps): React.ReactElement<any>;
     router?: CuriRouter;
     responsive?: boolean;
-}
-export interface CuriousRenderProps {
-    router: CuriRouter;
-    response: Response;
-    navigation: Navigation;
 }
 export interface CuriousState {
     response: Response;