+++ packages/react/types/interface.d.ts
@@ -1,9 +1,4 @@
-import { CuriRouter, Response, Navigation } from "@curi/core";
+import { ResponseHandlerProps } from "@curi/core";
 export interface CuriContext {
-    curi: CuriProps;
-}
-export interface CuriProps {
-    router: CuriRouter;
-    response: Response;
-    navigation: Navigation;
+    curi: ResponseHandlerProps;
 }