+++ packages/react/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* (Internal) Router's response handler receives `ResponseHandlerProps` object.
+
 ## 1.0.0-beta.19
 
 * Add a `<ResponsiveBase>` component. This is just a `<CuriBase>` wrapped in a `<Curious>`.